% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connector.R
\name{postgres_connector}
\alias{postgres_connector}
\title{postgres_connector}
\usage{
postgres_connector(
  host = Sys.getenv("POSTGRES_HOST"),
  port = Sys.getenv("POSTGRES_PORT"),
  user = Sys.getenv("POSTGRES_USER"),
  password = Sys.getenv("POSTGRES_PASSWORD"),
  dbname = Sys.getenv("POSTGRES_DB"),
  max_attempts = 2,
  verbose = TRUE
)
}
\arguments{
\item{host}{The hostame, Sys.getenv('POSTGRES_HOST')}

\item{port}{The port, Sys.getenv('POSTGRES_PORT')}

\item{user}{The port, Sys.getenv('POSTGRES_USER')}

\item{password}{The port, Sys.getenv('POSTGRES_PASSWORD')}

\item{dbname}{The dbname, Sys.getenv('POSTGRES_DB')}

\item{max_attempts}{Max retries if error}

\item{verbose}{Boolean, shut up or not}
}
\description{
postgres_connector
}
