% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_connector.R
\name{postgres_connector}
\alias{postgres_connector}
\title{postgres_connector}
\usage{
postgres_connector(
  host = Sys.getenv("POSTGRES_HOST"),
  port = Sys.getenv("POSTGRES_PORT"),
  user = Sys.getenv("POSTGRES_USER"),
  password = Sys.getenv("POSTGRES_PASSWORD"),
  dbname = Sys.getenv("POSTGRES_DB"),
  max_attempts = 2,
  verbose = TRUE
)
}
\description{
postgres_connector
}
